refactor(aggregated-assets): hide the fetch helpers (LIVE-35226)
Review feedback from @ysitbon on #20345: internal if nothing outside the
package uses it.

resolveBaseUrl and fetchAssetsPage have no consumer anywhere outside this
package, so they move to internals/requests.ts and leave the barrel.
buildAssetsQueryParams stays public — the LIVE-35224 characterization tests
import it through the dada-client shim, and those must not change.

collectAllByCategory duplicated resolveBaseUrl's getEnv branch inline; it now
calls it instead.

This is the cosmetic half of the point. Both helpers exist only because the
shared base query is configured with baseUrl: "", so every endpoint resolves
its own absolute url and the fan-out endpoint hand-rolls fetch — which is
also why assertDadaApiUrl is needed and why RTK's AbortSignal never reaches
the request. queryFn receives baseQuery as its fourth argument, so that is
fixable; both files now say so, and it is an acceptance criterion on
LIVE-35301 alongside the extraArgument migration.
